backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
openai==1.3.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
# PostgreSQL
psycopg2-binary==2.9.9
alembic==1.13.1
# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
email-validator>=2.0.0
# Rate limiting
slowapi==0.1.9
# Redis
redis==5.0.1
# Error tracking
sentry-sdk[fastapi]==1.40.0
# Similarity analysis (TF-IDF)
scikit-learn==1.3.2

